requirements.txt
ultralytics
opencv-python
pillow
pandas
numpy<2
matplotlib
seaborn
pyyaml
tqdm
fastapi
uvicorn
python-multipart
streamlit
